apr18: uint => int

+++ contracts/protocols/lending/aave3/Aave3PoolAdapterBase.sol
@@ -340,9 +340,9 @@ abstract contract Aave3PoolAdapterBase is IPoolAdapter, IPoolAdapterInitializer 
   }
 
   /// @notice Compute current cost of the money
-  function getAPR18() external view override returns (uint) {
+  function getAPR18() external view override returns (int) {
     Aave3DataTypes.ReserveData memory rb = _pool.getReserveData(borrowAsset);
-    return uint(rb.currentVariableBorrowRate) * 10**18 * 100 / 10**27;
+    return int(uint(rb.currentVariableBorrowRate) * 10**18 * 100 / 10**27);
   }
